:pencil2: adicionar verificação de existência no DELETE de Turma

+++ src/main/java/com/unilopers/mini_erp_escolar/controller/TurmaController.java
@@ -62,6 +62,9 @@ public class TurmaController {
     //DELETE
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable Long id) {
+        if (!turmaRepository.existsById(id)) {
+            throw new RuntimeException("Turma não encontrada");
+        }
         turmaRepository.deleteById(id);
         return ResponseEntity.noContent().build();
     }
